requirements.txt
torch==2.1.0
torchvision==0.16.0
accelerate==1.12.0
av==11.0.0
decord==0.6.0
diffusers==0.27.0
einops==0.8.1
fastapi==0.117.1
huggingface-hub==0.25.1
markdown2==2.5.4
mediapipe==0.10.11
numpy==1.26.3
omegaconf==2.3.0
onnx==1.17.0
onnx_graphsurgeon==0.5.8
onnx_simplifier==0.4.36
opencv_contrib_python==4.11.0.86
opencv_python_headless==4.10.0.84
Pillow==11.3.0
polygraphy==0.49.26
pycuda==2024.1.2
pydantic==2.12.5
safetensors==0.7.0
scikit-image==0.22.0
starlette==0.48.0
tensorrt==10.1.0
tqdm==4.67.1
transformers==4.36.2
uvicorn==0.37.0
xformers==0.0.22.post7
peft==0.10.0
python-multipart==0.0.20
uvicorn==0.37.0
onnxruntime==1.16.3
colored==2.3.1